Split kafka implementation into multiple files and use buffered channel for internal message passing.
